requirements.txt
streamlit
torch
torchaudio
pyannote.audio
speechbrain
ffmpeg-python
numpy
scipy
huggingface_hub
tqdm
